+++ NEWS.md
@@ -1,5 +1,6 @@
 # zoomGroupStats 0.1.9000
-* Added `imageDetectTime()` function to extract the recording time from a still frame of a Zoom video
+* Added `getVideoDateTimeMeta()` function to extract the creation timestamp for the recording from the video metadata
+* Added `imageDetectTimeStamp()` function to extract the recording time from a still frame of a Zoom video
 
 # zoomGroupStats 0.1.0
 * First full release of package